move selector to separate file
