+++ src/components/MoneyRequestReportView/MoneyRequestReportActionsList.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable rulesdir/prefer-early-return */
 import {useIsFocused, useRoute} from '@react-navigation/native';
-import {stableReportSelector} from '@selectors/Report';
+import stableReportSelector from '@selectors/stableReportSelector';
 import isEmpty from 'lodash/isEmpty';
 import React, {useCallback, useContext, useEffect, useLayoutEffect, useMemo, useRef, useState} from 'react';
 import type {LayoutChangeEvent, ListRenderItemInfo, NativeScrollEvent, NativeSyntheticEvent} from 'react-native';
