+++ src/selectors/Report.ts
@@ -47,88 +47,4 @@ function openExpenseReportIDsSelector(reports: OnyxCollection<Report>): OpenExpe
     return openExpenseReportIDMap;
 }
 
-/**
- * Stable `Report` projection for the `ReportActionItem` subtree.
- *
- * Commented-out `last*` fields are deliberately omitted: they are "heartbeat" fields that
- * update on routine activity (incoming/outgoing messages, read receipts, etc.) and would
- * invalidate the memo on every chat heartbeat without any of the item subtree actually
- * reading them. Everything else stays in the projection so consumer utils that read
- * Report fields keep working without surprise `undefined`s.
- */
-function stableReportSelector(report: OnyxEntry<Report>): OnyxEntry<Report> {
-    if (!report?.reportID) {
-        return undefined;
-    }
-    return {
-        reportID: report.reportID,
-        avatarUrl: report.avatarUrl,
-        created: report.created,
-        submitted: report.submitted,
-        approved: report.approved,
-        chatType: report.chatType,
-        hasOutstandingChildRequest: report.hasOutstandingChildRequest,
-        hasOutstandingChildTask: report.hasOutstandingChildTask,
-        isOwnPolicyExpenseChat: report.isOwnPolicyExpenseChat,
-        isPinned: report.isPinned,
-        // lastMessageText: report.lastMessageText,
-        // lastVisibleActionCreated: report.lastVisibleActionCreated,
-        // lastReadTime: report.lastReadTime,
-        // lastReadSequenceNumber: report.lastReadSequenceNumber,
-        // lastMentionedTime: report.lastMentionedTime,
-        policyAvatar: report.policyAvatar,
-        policyName: report.policyName,
-        oldPolicyName: report.oldPolicyName,
-        hasParentAccess: report.hasParentAccess,
-        description: report.description,
-        isDeletedParentAction: report.isDeletedParentAction,
-        policyID: report.policyID,
-        reportName: report.reportName,
-        chatReportID: report.chatReportID,
-        stateNum: report.stateNum,
-        statusNum: report.statusNum,
-        writeCapability: report.writeCapability,
-        type: report.type,
-        visibility: report.visibility,
-        invoiceReceiver: report.invoiceReceiver,
-        transactionCount: report.transactionCount,
-        parentReportID: report.parentReportID,
-        parentReportActionID: report.parentReportActionID,
-        // Coerce sentinel `0` to `undefined`. The backend ships `managerID: 0` on chat reports
-        // without a manager, and a later push removes the key entirely; treating both as
-        // `undefined` keeps the projection stable through that reconciliation.
-        // eslint-disable-next-line @typescript-eslint/prefer-nullish-coalescing
-        managerID: report.managerID || undefined,
-        // lastVisibleActionLastModified: report.lastVisibleActionLastModified,
-        // lastMessageHtml: report.lastMessageHtml,
-        // lastActorAccountID: report.lastActorAccountID,
-        // lastActionType: report.lastActionType,
-        ownerAccountID: report.ownerAccountID,
-        participants: report.participants,
-        total: report.total,
-        unheldTotal: report.unheldTotal,
-        unheldNonReimbursableTotal: report.unheldNonReimbursableTotal,
-        currency: report.currency,
-        errorFields: report.errorFields,
-        errors: report.errors,
-        isWaitingOnBankAccount: report.isWaitingOnBankAccount,
-        isCancelledIOU: report.isCancelledIOU,
-        hasReportBeenRetracted: report.hasReportBeenRetracted,
-        hasReportBeenReopened: report.hasReportBeenReopened,
-        isExportedToIntegration: report.isExportedToIntegration,
-        hasExportError: report.hasExportError,
-        iouReportID: report.iouReportID,
-        preexistingReportID: report.preexistingReportID,
-        nonReimbursableTotal: report.nonReimbursableTotal,
-        privateNotes: report.privateNotes,
-        fieldList: report.fieldList,
-        permissions: report.permissions,
-        tripData: report.tripData,
-        welcomeMessage: report.welcomeMessage,
-        nextStep: report.nextStep,
-        pendingAction: report.pendingAction,
-        pendingFields: report.pendingFields,
-    } as Report;
-}
-
-export {getArchiveReason, getReportChatType, getReportOwnerAccountID, getReportPolicyID, openExpenseReportIDsSelector, stableReportSelector};
+export {getArchiveReason, getReportChatType, getReportOwnerAccountID, getReportPolicyID, openExpenseReportIDsSelector};
